+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added lesson title into section above lesson content.
 - Improved styles for disabled buttons.
 
+## [2.6.4]
+- Added automatic switching to the section with search keywords if the first section does not contain any
+- Use "span" html tag instead of "marker" for highlighted paragraph
+- Fixed allowed values for "align" property in ImageBulletItem component
+- Changed default parse mode for search from "Single phrase" to "Multiple words" for better searchability
+
 ## [2.6.3]
 - Changed range of weights on page where courses are sorted to support ordering of more than 40 courses per category.